+++ overview-repo-mongo/src/test/java/cz/etn/overview/mongo/MongoVoucherMapper.java
@@ -76,7 +76,7 @@ public class MongoVoucherMapper extends DynamicEntityMapper<Voucher, Object> {
     }
 
     @Override
-    public String getDataSet() {
+    public String getTableName() {
         return COLLECTION_NAME;
     }