+++ overview-repo-mongo/src/test/java/cz/etn/overview/repo/VoucherMapper.java
@@ -76,7 +76,7 @@ public class VoucherMapper extends DynamicEntityMapper<Voucher, Object> {
 	}
 
 	@Override
-	public String getDataSet() {
+	public String getTableName() {
 		return DB_TABLE_NAME;
 	}