+++ overview-repo-sql/src/test/java/cz/etn/overview/sql/mapper/MySqlSchemaBuilderTest.java
@@ -12,9 +12,9 @@ public class MySqlSchemaBuilderTest {
     @Test
     public void composeCreateTableSQL() {
         VoucherMapper mapper = VoucherMapper.getInstance().getInstance();
-        String sql = new MySqlSchemaBuilder().composeCreateTableSQL(mapper.getDataSet(), mapper.getAttributes());
+        String sql = new MySqlSchemaBuilder().composeCreateTableSQL(mapper.getTableName(), mapper.getAttributes());
         System.out.println(sql);
-        assertTrue("Generated CREATE TABLE command is not as expected", sql.contains("CREATE TABLE IF NOT EXISTS `" + mapper.getDataSet() + "`"));
+        assertTrue("Generated CREATE TABLE command is not as expected", sql.contains("CREATE TABLE IF NOT EXISTS `" + mapper.getTableName() + "`"));
         assertTrue("Generated CREATE TABLE command does not contain code attribute", sql.contains("`code` VARCHAR(20) NOT NULL"));
     }